🚀 Rename `api` module to `client`, update related imports and classes for consistency
