+++ README.md
@@ -18,7 +18,7 @@ cp .env.template .env
 
 ```
 ├── app/                  # 애플리케이션 소스 코드
-    ├── api/              # API 엔드포인트 및 관련 로직 
+    ├── client/           # API 클라이언트 및 관련 로직 
     ├── config/           # 데이터베이스, 로그 등 설정 
     ├── dto/              # 데이터 전송 객체 
     ├── entity/           # 데이터베이스 테이블과 매핑되는 엔티티
