+++ app/task/tasks.py
@@ -8,8 +8,8 @@ from typing import Any
 import pandas as pd
 from celery_once import QueueOnce
 
-from app.api.naver_shopping_api import NaverShoppingApi
-from app.api.slack_client import SlackClient
+from app.client.naver_shopping_client import NaverShoppingClient
+from app.client.slack_client import SlackClient
 from app.config.settings import get_settings
 from app.enum.channel_enum import ChannelEnum
 from app.repository.model.search_conditions import ScrapedProductSearchCondition
@@ -25,17 +25,17 @@ settings = get_settings()
 
 @celery_app.task(base=QueueOnce, once={"graceful": True, "timeout": 60 * 5})
 def scrape_naver_shopping_task():
-    naver_shopping_api = NaverShoppingApi()
+    naver_shopping_client = NaverShoppingClient()
     keyword_service = KeywordService()
     scraped_product_service = ScrapedProductService()
 
     scrape_products_by_keywords(
-        naver_shopping_api,
+        naver_shopping_client,
         keyword_service,
         scraped_product_service,
     )
     scrape_tracking_required_products(
-        naver_shopping_api,
+        naver_shopping_client,
         scraped_product_service,
     )
     excel_filepath = create_excel_from_scraped_products(scraped_product_service)
@@ -98,7 +98,7 @@ def flatten_scraped_product_details(
 
 
 def scrape_products_by_keywords(
-    naver_shopping_api: NaverShoppingApi,
+    naver_shopping_client: NaverShoppingClient,
     keyword_service: KeywordService,
     scraped_product_service: ScrapedProductService,
 ):
@@ -109,7 +109,7 @@ def scrape_products_by_keywords(
         return
 
     for keyword in keywords:
-        searched_items = naver_shopping_api.search_with_all_pages(keyword.word)
+        searched_items = naver_shopping_client.search_with_all_pages(keyword.word)
         scraped_product_service.save_naver_shopping_search_result(
             searched_items,
             keyword.id,
@@ -121,7 +121,7 @@ def scrape_products_by_keywords(
 
 
 def scrape_tracking_required_products(
-    naver_shopping_api: NaverShoppingApi,
+    naver_shopping_client: NaverShoppingClient,
     scraped_product_service: ScrapedProductService,
 ):
     logger.info("[SCRAPE_TRACKING_REQUIRED_PRODUCTS] 🚀 추가 트래킹 필요한 상품정보 수집 시작 🚀")
@@ -130,7 +130,7 @@ def scrape_tracking_required_products(
         ChannelEnum.NAVER_SHOPPING
     )
     for product in scraped_products_tracking_required:
-        searched_items = naver_shopping_api.search_with_all_pages(product.name)
+        searched_items = naver_shopping_client.search_with_all_pages(product.name)
         scraped_product_service.save_naver_shopping_search_result(
             searched_items,
             product.keyword.id,
